+++ ent/contest/where.go
@@ -66,12 +66,12 @@ func Description(v string) predicate.Contest {
 }
 
 // Status applies equality check predicate on the "status" field. It's identical to StatusEQ.
-func Status(v int) predicate.Contest {
+func Status(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldEQ(FieldStatus, v))
 }
 
 // Type applies equality check predicate on the "type" field. It's identical to TypeEQ.
-func Type(v int) predicate.Contest {
+func Type(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldEQ(FieldType, v))
 }
 
@@ -91,7 +91,7 @@ func Language(v string) predicate.Contest {
 }
 
 // ExtraTime applies equality check predicate on the "extra_time" field. It's identical to ExtraTimeEQ.
-func ExtraTime(v int) predicate.Contest {
+func ExtraTime(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldEQ(FieldExtraTime, v))
 }
 
@@ -231,82 +231,82 @@ func DescriptionContainsFold(v string) predicate.Contest {
 }
 
 // StatusEQ applies the EQ predicate on the "status" field.
-func StatusEQ(v int) predicate.Contest {
+func StatusEQ(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldEQ(FieldStatus, v))
 }
 
 // StatusNEQ applies the NEQ predicate on the "status" field.
-func StatusNEQ(v int) predicate.Contest {
+func StatusNEQ(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldNEQ(FieldStatus, v))
 }
 
 // StatusIn applies the In predicate on the "status" field.
-func StatusIn(vs ...int) predicate.Contest {
+func StatusIn(vs ...int16) predicate.Contest {
 	return predicate.Contest(sql.FieldIn(FieldStatus, vs...))
 }
 
 // StatusNotIn applies the NotIn predicate on the "status" field.
-func StatusNotIn(vs ...int) predicate.Contest {
+func StatusNotIn(vs ...int16) predicate.Contest {
 	return predicate.Contest(sql.FieldNotIn(FieldStatus, vs...))
 }
 
 // StatusGT applies the GT predicate on the "status" field.
-func StatusGT(v int) predicate.Contest {
+func StatusGT(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldGT(FieldStatus, v))
 }
 
 // StatusGTE applies the GTE predicate on the "status" field.
-func StatusGTE(v int) predicate.Contest {
+func StatusGTE(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldGTE(FieldStatus, v))
 }
 
 // StatusLT applies the LT predicate on the "status" field.
-func StatusLT(v int) predicate.Contest {
+func StatusLT(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldLT(FieldStatus, v))
 }
 
 // StatusLTE applies the LTE predicate on the "status" field.
-func StatusLTE(v int) predicate.Contest {
+func StatusLTE(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldLTE(FieldStatus, v))
 }
 
 // TypeEQ applies the EQ predicate on the "type" field.
-func TypeEQ(v int) predicate.Contest {
+func TypeEQ(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldEQ(FieldType, v))
 }
 
 // TypeNEQ applies the NEQ predicate on the "type" field.
-func TypeNEQ(v int) predicate.Contest {
+func TypeNEQ(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldNEQ(FieldType, v))
 }
 
 // TypeIn applies the In predicate on the "type" field.
-func TypeIn(vs ...int) predicate.Contest {
+func TypeIn(vs ...int16) predicate.Contest {
 	return predicate.Contest(sql.FieldIn(FieldType, vs...))
 }
 
 // TypeNotIn applies the NotIn predicate on the "type" field.
-func TypeNotIn(vs ...int) predicate.Contest {
+func TypeNotIn(vs ...int16) predicate.Contest {
 	return predicate.Contest(sql.FieldNotIn(FieldType, vs...))
 }
 
 // TypeGT applies the GT predicate on the "type" field.
-func TypeGT(v int) predicate.Contest {
+func TypeGT(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldGT(FieldType, v))
 }
 
 // TypeGTE applies the GTE predicate on the "type" field.
-func TypeGTE(v int) predicate.Contest {
+func TypeGTE(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldGTE(FieldType, v))
 }
 
 // TypeLT applies the LT predicate on the "type" field.
-func TypeLT(v int) predicate.Contest {
+func TypeLT(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldLT(FieldType, v))
 }
 
 // TypeLTE applies the LTE predicate on the "type" field.
-func TypeLTE(v int) predicate.Contest {
+func TypeLTE(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldLTE(FieldType, v))
 }
 
@@ -456,42 +456,42 @@ func LanguageContainsFold(v string) predicate.Contest {
 }
 
 // ExtraTimeEQ applies the EQ predicate on the "extra_time" field.
-func ExtraTimeEQ(v int) predicate.Contest {
+func ExtraTimeEQ(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldEQ(FieldExtraTime, v))
 }
 
 // ExtraTimeNEQ applies the NEQ predicate on the "extra_time" field.
-func ExtraTimeNEQ(v int) predicate.Contest {
+func ExtraTimeNEQ(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldNEQ(FieldExtraTime, v))
 }
 
 // ExtraTimeIn applies the In predicate on the "extra_time" field.
-func ExtraTimeIn(vs ...int) predicate.Contest {
+func ExtraTimeIn(vs ...int16) predicate.Contest {
 	return predicate.Contest(sql.FieldIn(FieldExtraTime, vs...))
 }
 
 // ExtraTimeNotIn applies the NotIn predicate on the "extra_time" field.
-func ExtraTimeNotIn(vs ...int) predicate.Contest {
+func ExtraTimeNotIn(vs ...int16) predicate.Contest {
 	return predicate.Contest(sql.FieldNotIn(FieldExtraTime, vs...))
 }
 
 // ExtraTimeGT applies the GT predicate on the "extra_time" field.
-func ExtraTimeGT(v int) predicate.Contest {
+func ExtraTimeGT(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldGT(FieldExtraTime, v))
 }
 
 // ExtraTimeGTE applies the GTE predicate on the "extra_time" field.
-func ExtraTimeGTE(v int) predicate.Contest {
+func ExtraTimeGTE(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldGTE(FieldExtraTime, v))
 }
 
 // ExtraTimeLT applies the LT predicate on the "extra_time" field.
-func ExtraTimeLT(v int) predicate.Contest {
+func ExtraTimeLT(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldLT(FieldExtraTime, v))
 }
 
 // ExtraTimeLTE applies the LTE predicate on the "extra_time" field.
-func ExtraTimeLTE(v int) predicate.Contest {
+func ExtraTimeLTE(v int16) predicate.Contest {
 	return predicate.Contest(sql.FieldLTE(FieldExtraTime, v))
 }
